Completing merge of SqlClientFactory changes
