+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/SqlClientFactory.cs
@@ -3,14 +3,19 @@
 // See the LICENSE file in the project root for more information.
 
 using Microsoft.Data.Sql;
+using System;
 using System.Data.Common;
 using System.Security.Permissions;
 using System.Security;
+using Microsoft.Data.Common;
 
 namespace Microsoft.Data.SqlClient
 {
     /// <include file='../../../../../../doc/snippets/Microsoft.Data.SqlClient/SqlClientFactory.xml' path='docs/members[@name="SqlClientFactory"]/SqlClientFactory/*'/>
-    public sealed partial class SqlClientFactory : DbProviderFactory
+    public sealed class SqlClientFactory : DbProviderFactory
+#if NETFRAMEWORK
+        , IServiceProvider
+#endif
     {
 
         /// <include file='../../../../../../doc/snippets/Microsoft.Data.SqlClient/SqlClientFactory.xml' path='docs/members[@name="SqlClientFactory"]/Instance/*'/>
@@ -64,6 +69,21 @@ namespace Microsoft.Data.SqlClient
         public override CodeAccessPermission CreatePermission(PermissionState state)
         {
             return new SqlClientPermission(state);
+        }
+
+        /// <summary>
+        /// Extension mechanism for additional services; currently the only service
+        /// supported is the DbProviderServices
+        /// </summary>
+        /// <returns>requested service provider or null.</returns>
+        object IServiceProvider.GetService(Type serviceType)
+        {
+            object result = null;
+            if (serviceType == GreenMethods.SystemDataCommonDbProviderServices_Type)
+            {
+                result = GreenMethods.MicrosoftDataSqlClientSqlProviderServices_Instance();
+            }
+            return result;
         }
 #endif
